Move test fixture to conftest.py
